Refactor tests and README
